+++ core/mc-semantic-ui-core/src/main/java/ai/mindconnect/ui/model/UiNode.java
@@ -49,6 +49,7 @@ import java.util.Map;
     @JsonSubTypes.Type(value = UiUpload.class,    name = "upload"),
     @JsonSubTypes.Type(value = UiLink.class,      name = "link"),
     @JsonSubTypes.Type(value = UiDialog.class,    name = "dialog"),
+    @JsonSubTypes.Type(value = UiDrawer.class,    name = "drawer"),
     @JsonSubTypes.Type(value = UiPage.class,      name = "page"),
 })
 @JsonInclude(JsonInclude.Include.NON_NULL)